separate story output tests
